Extract `NotificationStreamAssembler` into a file
